Refactor out most of the "RPI" into a crate called "mtarp" -- Multi-Tenant Arpeggio
mtarp contains all the code that is fundamentally about managing multiple games and multiple users, and the storage of those games and users.

RPI is now focused specifically on the web server part, including the google-based authentication layer and the long-poll PingService, and of course all the HTTP routing.
