+++ mtarp/Cargo.toml
@@ -1,13 +1,12 @@
 [package]
-name = "rpi"
-description = "Arpeggio's Remote Programming Interface (web-based) -- in other words, the web app that hosts Arpeggio games."
+name = "mtarp"
+description = "Multi-Tenant Arpeggio -- the types & behavior for managing a service that runs multiple games simultaneously."
 version = "0.1.0"
 authors = ["Christopher Armstrong"]
 edition = "2021"
 
 [dependencies]
 async-trait = "0.1.73"
-axum = {version = "0.6.20"}
 foldertree = {version = "*", path = "../foldertree"}
 anyhow = {version = "1.0.75", features = ["backtrace"]}
 thiserror = "1.0.48"
@@ -16,19 +15,12 @@ http = "0.2.0"
 tracing = { version = "0.1.37", features = ["attributes"]}
 tracing-subscriber = {version = "0.3.17", features = ["env-filter", "json"]}
 arpeggio = {version = "*", path = "../arpeggio"}
-mtarp = {version = "*", path = "../mtarp"}
 serde = "1.0"
 serde_derive = "1.0"
 serde_json = "1.0"
 serde_yaml = "0.8"
 clap = {version = "4.4.6", features = ["derive"]}
-tokio = {version = "1.32", features = ["time", "sync", "rt", "macros", "rt-multi-thread"]}
-tower-http = {version = "0.4.4", features = ["cors", "trace"]}
 ts-rs = {version = "7.0.0", features = ["serde-compat", "uuid-impl"]}
+# Why do I have this uom dependency here?
 uom = {version = "0.35", default_features = false}
 uuid = { version = "1.1.2", features = ["v4", "serde"] }
-google-cloud-storage = "0.13.1"
-# Unfortunately this google-signin package seems pretty unmaintained.
-# I need to use the git version because the support for local verification of ID tokens is only in
-# an unreleased version.
-google-signin = { git = "https://github.com/wyyerd/google-signin-rs" }
