--- rpi/buggy-game.yaml
@@ -1,3 +1,0 @@
----
-content: "Foo\
-    Bar"
