--- rpi/modules/fantasy-sample.yaml
@@ -1,168 +1,0 @@
----
-current_game:
-  current_combat: ~
-  abilities:
-    "00000000-0000-0000-0000-000000000007":
-      id: "00000000-0000-0000-0000-000000000007"
-      name: Heal
-      cost: 8
-      action:
-        Creature:
-          effect:
-            Heal:
-              Expr:
-                num: 2
-                size: 2
-          target:
-            Range: 500
-      usable_ooc: true
-    "00000000-0000-0000-0000-000000000001":
-      id: "00000000-0000-0000-0000-000000000001"
-      name: Piercing Shot
-      cost: 8
-      action:
-        Creature:
-          effect:
-            Damage:
-              Expr:
-                num: 1
-                size: 3
-          target:
-            LineFromActor:
-              distance: 1000
-      usable_ooc: false
-    "00000000-0000-0000-0000-000000000002":
-      id: "00000000-0000-0000-0000-000000000002"
-      name: Shoot
-      cost: 8
-      action:
-        Creature:
-          effect:
-            Damage:
-              Expr:
-                num: 2
-                size: 2
-          target:
-            Range: 1000
-      usable_ooc: false
-    "00000000-0000-0000-0000-000000000000":
-      id: "00000000-0000-0000-0000-000000000000"
-      name: Caltrops
-      cost: 8
-      action:
-        SceneVolume:
-          effect:
-            CreateVolumeCondition:
-              duration: Interminate
-              condition:
-                RecurringEffect:
-                  Damage:
-                    Expr:
-                      num: 1
-                      size: 3
-          target:
-            RangedVolume:
-              volume:
-                Sphere: 200
-              range: 1000
-      usable_ooc: false
-    "00000000-0000-0000-0000-000000000006":
-      id: "00000000-0000-0000-0000-000000000006"
-      name: Stab
-      cost: 8
-      action:
-        Creature:
-          effect:
-            Damage:
-              Expr:
-                num: 3
-                size: 1
-          target: Melee
-      usable_ooc: false
-    "00000000-0000-0000-0000-000000000003":
-      id: "00000000-0000-0000-0000-000000000003"
-      name: Fireball
-      cost: 10
-      action:
-        Creature:
-          effect:
-            Damage:
-              Expr:
-                num: 2
-                size: 5
-          target:
-            AllCreaturesInVolumeInRange:
-              volume:
-                Sphere: 500
-              range: 1000
-      usable_ooc: false
-  creatures: {}
-  classes:
-    "00000000-0000-0000-0000-000000000002":
-      id: "00000000-0000-0000-0000-000000000002"
-      name: Archer
-      abilities:
-        - "00000000-0000-0000-0000-000000000002"
-        - "00000000-0000-0000-0000-000000000001"
-        - "00000000-0000-0000-0000-000000000000"
-      conditions:
-        - RecurringEffect:
-            GenerateEnergy: 10
-      color: darkgreen
-    "00000000-0000-0000-0000-000000000001":
-      id: "00000000-0000-0000-0000-000000000001"
-      name: Mage
-      abilities:
-        - "00000000-0000-0000-0000-000000000003"
-      conditions: []
-      color: purple
-    "00000000-0000-0000-0000-000000000004":
-      id: "00000000-0000-0000-0000-000000000004"
-      name: Rogue
-      abilities:
-        - "00000000-0000-0000-0000-000000000006"
-      conditions:
-        - RecurringEffect:
-            GenerateEnergy: 10
-      color: gray
-    "00000000-0000-0000-0000-000000000003":
-      id: "00000000-0000-0000-0000-000000000003"
-      name: Healer
-      abilities:
-        - "00000000-0000-0000-0000-000000000007"
-      conditions:
-        - RecurringEffect:
-            GenerateEnergy: 10
-      color: lightgreen
-    "00000000-0000-0000-0000-000000000000":
-      id: "00000000-0000-0000-0000-000000000000"
-      name: Knight
-      abilities: []
-      conditions: []
-      color: red
-  tile_system: Realistic
-  scenes: {}
-  items: {}
-  campaign:
-    data:
-      scenes: []
-      creatures: []
-      notes: {}
-      items: []
-      abilities:
-        - "00000000-0000-0000-0000-000000000007"
-        - "00000000-0000-0000-0000-000000000001"
-        - "00000000-0000-0000-0000-000000000000"
-        - "00000000-0000-0000-0000-000000000006"
-        - "00000000-0000-0000-0000-000000000002"
-        - "00000000-0000-0000-0000-000000000003"
-      classes:
-        - "00000000-0000-0000-0000-000000000001"
-        - "00000000-0000-0000-0000-000000000002"
-        - "00000000-0000-0000-0000-000000000004"
-        - "00000000-0000-0000-0000-000000000000"
-        - "00000000-0000-0000-0000-000000000003"
-    children: {}
-  players: {}
-  active_scene: ~
-snapshots: []
