+++ rpi/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod actor;
-pub mod storage;
-pub mod types;
 pub mod web;
+pub mod waiters;
+pub mod authn;
