+++ worker/Cargo.toml
@@ -16,6 +16,7 @@ serde_json = "1.0"
 serde-wasm-bindgen = "0.6.0"
 console_error_panic_hook = "0.1.6"
 arpeggio = {version = "*", path = "../arpeggio"}
+mtarp = {version = "*", path = "../mtarp"}
 # arpeggio depend on rand, which depends on getrandom, which by default does not work on WASM. We need to enable the "js" feature.
 getrandom = {version = "0.2.9", features = ["js"]}
